+++ app/Http/Controllers/RolesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Role;
+use App\User;
 use App\Permission;
 use Illuminate\Http\Request;
 
@@ -10,22 +11,17 @@ class RolesController extends Controller
 {
     public function index()
     {
-        return view('admin.sites.roles.show', [
+        $this->authorize('admin', User::class);
+
+        return view('admin.sites.roles.role_index', [
             'roles' => Role::with('permissions')->get(),
             'permissions' => Permission::where('name', '!=', 'admin')->get(),
         ]);
     }
 
-    public function add_permission(Request $request)
-    {
-        $permission = Permission::find($request->permission);
-        Role::find($request->role_id)->permissions()->save($permission);
-
-        return back();
-    }
-
     public function store(Request $request)
     {
+        $this->authorize('admin', User::class);
         $this->validate($request, [
             'name' => 'required|string|unique:roles|max:20',
             'label' => 'required|string|max:50',
@@ -44,6 +40,7 @@ class RolesController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('admin', User::class);
         Role::destroy($id);
 
         return back();
@@ -51,6 +48,7 @@ class RolesController extends Controller
 
     public function destroy_permission(Role $role, $permission_id)
     {
+        $this->authorize('admin', User::class);
         $role->permissions()->detach($permission_id);
 
         return back();
